requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
httpx
python-dotenv
sqlalchemy
aiosqlite
